Scripts/nba_predictor/requirements.txt
functions-framework==3.*
firebase-admin==6.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
pandas
joblib
scikit-learn